requirements.txt
requests>=2.31.0
duckdb>=0.9.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
streamlit>=1.29.0
plotly>=5.17.0
matplotlib>=3.8.0
seaborn>=0.13.0
openpyxl>=3.1.0
xlrd>=2.0.0 